feat: ping project on {r+-, d+-, p=} (#41)
`Project.ping!/1` sends a websocket notification to users browsing the project page, making it show "This page is out of date. Reload" at the top of the page. The `r+`, `r-`, `d+`, `d-` and `p=` commands change the data displayed on the project page but did not trigger a ping; this PR rectifies that.